updated readme.md file, added protection against multiple instances, actually centralized the parsehotkeystring function to hotkey_utils
